+++ src/gdconvert/conversion_scalar.hpp
@@ -1,0 +1,11 @@
+#ifndef CONVERSION_SCALAR_HPP
+#define CONVERSION_SCALAR_HPP
+
+#include "vatensor/varray.hpp"            // for VScalar
+#include "godot_cpp/variant/variant.hpp"  // for Variant
+
+using namespace godot;
+
+va::VScalar variant_to_vscalar(const Variant& variant);
+
+#endif //CONVERSION_SCALAR_HPP